requirements.txt
SpeechRecognition==3.10.1
pyttsx3==2.90
google-genai==0.3.0
python-dotenv==1.0.1
pydantic==2.6.4
PyAudio==0.2.14
openai>=1.12.0
